refactor(ui): split useGameSounds into sounds/ submodule
useGameSounds.ts went 523 → 74 lines. Three backends (Tauri rodio,
Web Audio, HTML Audio), the manager (init + dispatch + error reporting),
and the global enabled state each live in their own file under
packages/ui/src/sounds/. The hook is now a thin debounce + variant
rotation wrapper around playSoundInternal.

Public exports (useGameSounds, setSoundInitErrorHandler, SoundInitError)
re-exported from the same path so consumers don't change.
